Module aurae_proto::runtime

Modules

cell_service_client

Generated client implementations.

cell_service_server

Generated server implementations.

instance_service_client

Generated client implementations.

instance_service_server

Generated server implementations.

pod_service_client

Generated client implementations.

pod_service_server

Generated server implementations.

spawn_service_client

Generated client implementations.

spawn_service_server

Generated server implementations.

Structs

Cell

/ An isolation resource used to divide a system into smaller resource / boundaries.

CellServiceAllocateRequest

/ An Aurae cell is a name given to Linux control groups (cgroups) that also include / a name, and special pre-exec functionality that is executed from within the same context / as any executables scheduled. / / A cell must be allocated for every executable scheduled. A cell defines the resource / constraints of the system to allocate for an arbitrary use case.

CellServiceAllocateResponse

/ The response after a cell has been allocated.

CellServiceFreeRequest

/ Used to remove or free a cell after it has been allocated.

CellServiceFreeResponse

/ Response after removing or freeing a cell.

CellServiceStartRequest

/ A request for starting an executable inside of a Cell. / / This is the lowest level of raw executive functionality. / Here you can define shell commands, and meta information about the command. / An executable is started synchronously.

CellServiceStartResponse

/ The response after starting an executable within a Cell.

CellServiceStopRequest

/ Request to stop an executable at runtime.

CellServiceStopResponse

Container

Executable

/ The most primitive workload in Aurae, a standard executable process.

Pod

/ OCI image represents a filesystem bundle on disk using familiar parlance. / / OCI Filesystem Bundle: https://github.com/opencontainers/runtime-spec/blob/main/bundle.md#filesystem-bundle

PodServiceAllocateRequest

/ The request to allocate a Pod.

PodServiceAllocateResponse

PodServiceFreeRequest

PodServiceFreeResponse

PodServiceStartRequest

PodServiceStartResponse

PodServiceStopRequest

PodServiceStopResponse

Constants

FILE_DESCRIPTOR_SET

Encoded file descriptor set for the aurae.runtime.v0 package